requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.14
django-ckeditor==6.7.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
setuptools==71.1.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
